+++ pkgs/native_assets_builder/test_data/package_with_metadata/build.dart
@@ -2,19 +2,15 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// TODO(https://github.com/dart-lang/native/issues/882): A new version of
-// `package:native_assets_cli` needs to be published first.
-// ignore_for_file: deprecated_member_use
-
 import 'package:native_assets_cli/native_assets_cli.dart';
 
 void main(List<String> args) async {
   final buildConfig = await BuildConfig.fromArgs(args);
   final buildOutput = BuildOutput(
-    metadata: Metadata({
+    metadata: {
       'some_key': 'some_value',
       'some_int': 3,
-    }),
+    },
   );
   await buildOutput.writeToFile(outDir: buildConfig.outDir);
 }