+++ pkgs/native_assets_cli/test/model/build_config_test.dart
@@ -90,7 +90,6 @@ void main() async {
         true);
     expect(config1.cCompiler != config2.cCompiler, true);
     expect(config1.linkModePreference, config2.linkModePreference);
-    expect(config1.dependencyMetadata, config2.dependencyMetadata);
   });
 
   test('BuildConfig fromConfig', () {